WIP: first contact with Stripe achieved
This point of all this work has been to reach this point: I've been able
to install and configure everything so that I can see that stripe.js
object show up. At this point I think I should throw this away and start
over on the upgrade page. The goal then would be to actually create a
customer object in Stripe that's connected to a User record in Postgres.