Fix flaky tests
These couple tests were added when I updated the pace logic and worked
just fine on certain days, but then on other days would never pass. What
I'm doing with this commit is locking time to a particular day and then
updating the tests to reflect what should be true on that day.