Improve Honeybadger config
This mostly boils down to specifying the api key and env using my
existing config pipeline rather than relying on ENV vars. I think this
is better because it allows me to set it both in Ruby land and also on
the client side. I have removed the associated vars from Heroku.

I also setup the JS config so that all client side exceptions will be
sent to Honeybadger as well.